Moved StyledComponents.tsx to the components subfolder. Now only screen logic is in the presentation folder, and the reusable code is in the common subfolder
