+++ AwesomeProject/src/presentation/common/components/StyledComponents.tsx
